fix(suite): on copy device display
