+++ packages/suite/src/utils/suite/deviceDisplay.ts
@@ -1,0 +1,11 @@
+import { ClipboardEvent } from 'react';
+
+export const handleOnCopy = (event: ClipboardEvent) => {
+    const selectedText = window.getSelection()?.toString();
+
+    if (selectedText) {
+        const processedText = selectedText.replace(/\s/g, '');
+        event?.nativeEvent?.clipboardData?.setData('text/plain', processedText);
+        event.preventDefault();
+    }
+};
